% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_path}
\alias{cache_path}
\title{Manage the rastermap file cache.}
\usage{
cache_path(...)
}
\description{
To avoid downloading the same maps again and again, rastermap locally
caches the you've downloaded. By default, they will be stored in a
session specific temporary directory which will be deleted each time you
quit R. If you want them to persist across sessions, set the global option
"rastermap.cache".
}
\examples{
dir(cache_path())
}
\keyword{internal}

