% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slippymaps.R
\name{fetch_region}
\alias{fetch_region}
\title{Fetch tile region.}
\usage{
fetch_region(lon, lat, provider, cache = TRUE, zoom = 10)
}
\arguments{
\item{lon,lat}{Longitude and latitiude ranges}

\item{provider}{A tile set provider. See \code{\link{providers}}
for possible options}

\item{cache}{If \code{TRUE}, uses cache as described in
\code{\link{cache_path}}}

\item{zoom}{Zoom level (0-18). Higher numbers produce more detailed
graphics, but will require more data to be downloaded.}
}
\description{
Given bounding box, find tiles that span region and stitch them together
into a single raster.
}
\examples{
houston <- fetch_region(c(-95.80204, -94.92313), c(29.38048, 30.14344),
  stamen("terrain"))
houston
plot(houston)
}

