% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/providers.R
\name{provider}
\alias{provider}
\title{Create a new provider.}
\usage{
provider(tile_f, attr_tile, attr_data = osm_license)
}
\arguments{
\item{tile_f}{A tile function has inputs x, y and z and returns
a url.}

\item{attr_title,attr_data}{Attribution information for rendered tiles
and underlying data.}
}
\description{
A provider wrappers a tile function with attribution details.
}
\keyword{internal}

