% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/providers.R
\name{providers}
\alias{cartodo}
\alias{cloudmade}
\alias{open_street_map}
\alias{providers}
\alias{stamen}
\title{Built-in tile set providers}
\usage{
stamen(style = "terrain")

open_street_map()

cloudmade(style, key)

cartodo()
}
\arguments{
\item{style}{Map style.}

\item{key}{api key}
}
\description{
Built-in tile set providers
}

